client/requirements.txt
paho-mqtt==1.6.1
psutil==5.9.4
python-dotenv==0.21.0
